+++ src/main/java/org/jabref/gui/collab/DatabaseChangeMonitor.java
@@ -1,14 +1,11 @@
 package org.jabref.gui.collab;
 
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.jabref.gui.util.BackgroundTask;
 import org.jabref.gui.util.TaskExecutor;
-import org.jabref.logic.util.io.FileUtil;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.util.FileUpdateListener;
 import org.jabref.model.util.FileUpdateMonitor;
@@ -22,7 +19,6 @@ public class DatabaseChangeMonitor implements FileUpdateListener {
     private final BibDatabaseContext database;
     private final FileUpdateMonitor fileMonitor;
     private final List<DatabaseChangeListener> listeners;
-    private Path referenceFile;
     private TaskExecutor taskExecutor;
 
     public DatabaseChangeMonitor(BibDatabaseContext database, FileUpdateMonitor fileMonitor, TaskExecutor taskExecutor) {
@@ -34,9 +30,6 @@ public class DatabaseChangeMonitor implements FileUpdateListener {
         this.database.getDatabasePath().ifPresent(path -> {
             try {
                 fileMonitor.addListenerForFile(path, this);
-                referenceFile = Files.createTempFile("jabref", ".bib");
-                referenceFile.toFile().deleteOnExit();
-                setAsReference(path);
             } catch (IOException e) {
                 LOGGER.error("Error while trying to monitor " + path, e);
             }
@@ -64,15 +57,4 @@ public class DatabaseChangeMonitor implements FileUpdateListener {
         database.getDatabasePath().ifPresent(file -> fileMonitor.removeListener(file, this));
     }
 
-    public void markExternalChangesAsResolved() {
-        markAsSaved();
-    }
-
-    public void markAsSaved() {
-        database.getDatabasePath().ifPresent(this::setAsReference);
-    }
-
-    private void setAsReference(Path file) {
-        FileUtil.copyFile(file, referenceFile, true);
-    }
 }